Some bugfixes
round should return float, but return Decimal, when passing Decimal to it, so I have to make it force float

+++ src/bitcoinrpc/proxy.py
@@ -39,7 +39,7 @@ HTTP_TIMEOUT = 30
 
 def EncodeDecimal(o):
     if isinstance(o, decimal.Decimal):
-        return round(o, 8)
+        return float(round(o, 8))
     raise TypeError(repr(o) + " is not JSON serializable")
 
 class JSONRPCException(Exception):
